+++ include/Game/Piece.h
@@ -22,7 +22,7 @@ class Piece
 
         std::string pgn_symbol() const;
         char fen_symbol() const;
-        std::string ascii_art(int row, int square_height) const;
+        std::string ascii_art(size_t row, size_t square_height) const;
         Color color() const;
 
         bool can_move(const Move* move) const;
@@ -74,7 +74,7 @@ class Piece
         void add_standard_legal_move(int file_step, int rank_step);
 
         std::vector<std::string> ascii_art_lines;
-        static int maximum_piece_height;
+        static size_t maximum_piece_height;
 
         void add_pawn_art();
         void add_rook_art();
