+++ include/Genes/Sphere_of_Influence_Gene.h
@@ -33,7 +33,7 @@ class Sphere_of_Influence_Gene : public Gene
 
         std::array<double, 8> scalar_cache;
 
-        static std::array<std::array<int, 64>, 64> king_distances;
+        static std::array<std::array<size_t, 64>, 64> king_distances;
         static bool king_distances_initialized;
 
         double score_board(const Board& board, Color perspective, const size_t depth) const override;
