+++ src/Genes/Genome.cpp
@@ -31,9 +31,9 @@
 
 #include "Exceptions/Genetic_AI_Creation_Error.h"
 
-size_t Genome::piece_strength_gene_index = -1;
-size_t Genome::look_ahead_gene_index = -1;
-size_t Genome::mutation_rate_gene_index = -1;
+size_t Genome::piece_strength_gene_index = size_t(-1);
+size_t Genome::look_ahead_gene_index = size_t(-1);
+size_t Genome::mutation_rate_gene_index = size_t(-1);
 
 //! Create a genome with default-constructed (neutral behavior) Genes.
 Genome::Genome()
